Rename ahead of publishing
