+++ pyGr/util.py
